Cambio de u/fixed32_t a u/fixed64_t: pasamos el rango a 64 bits y precisión de 8 a 16 bits. Cambios aplicados a todo el código menos al RenderSystem por restricciones de tinyPTC

+++ src/cmp/AI_Component.hpp
@@ -20,7 +20,7 @@ struct AI_Component : Component_t {
 
     std::vector<fixed_vec2> target_vec { fixed_vec2 { -300, -300 }, fixed_vec2 {  200, -300 }
                                        , fixed_vec2 {  200,  200 }, fixed_vec2 { -300,  200 } };
-    uint32_t target_index { 0u };
+    uint64_t target_index { 0u };
 
 };
